requirements.txt
docker
pyyaml
requests